Move a lot of business logic out of WithdrawalDialogForm, into withdraw.ts
