+++ src/components/TransferService/util.ts
@@ -1,4 +1,6 @@
 import React from "react"
+import { Memo } from "stellar-sdk"
+import { WithdrawalSuccessResponse } from "@satoshipay/stellar-sep-6"
 import { trackError } from "../../context/notifications"
 
 export function usePolling(pollIntervalMs: number) {
@@ -30,5 +32,25 @@ export function usePolling(pollIntervalMs: number) {
   return {
     start,
     stop
+  }
+}
+
+export function createMemo(withdrawalResponse: WithdrawalSuccessResponse): Memo | null {
+  const { memo, memo_type: type } = withdrawalResponse
+
+  if (!memo || !type) {
+    return null
+  }
+
+  switch (type) {
+    case "hash":
+      const hash = Buffer.from(memo, "base64")
+      return Memo.hash(hash.toString("hex"))
+    case "id":
+      return Memo.id(memo)
+    case "text":
+      return Memo.text(memo)
+    default:
+      return null
   }
 }
